[FramworkBundle][HttpKernel] fix KernelBrowser BC layer
